requirements.txt
Django==2.2.7
django-autoslug==1.9.6
django-jsonfield==1.3.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
Faker==2.0.4
sqlparse==0.3.0
